+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.16.5",
     "big-js": "^3.1.3",
     "big.js": "^6.2.1",
     "react": "^18.2.0",
@@ -42,7 +43,6 @@
     "@babel/eslint-parser": "^7.18.9",
     "@babel/plugin-syntax-jsx": "^7.18.6",
     "@babel/preset-react": "^7.18.6",
-    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.3.0",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
