+++ src/App.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import Calculator from './components/Calculator';
+import Home from './components/Home';
+import Quote from './components/Quote';
+import Header from './components/Header';
+
+function App() {
+  return (
+    <Router>
+      <Header />
+      <Routes>
+        <Route exact path="/" element={<Home />} />
+        <Route exact path="/Calculator" element={<Calculator />} />
+        <Route exact path="/Quote" element={<Quote />} />
+      </Routes>
+    </Router>
+  );
+}
+
+export default App;
